requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
joblib>=1.3.0
pytest>=8.0.0
# Local LLM for image prompt generation
# Note: Commented out for cloud deployment to stay under 1GB RAM limit.
# torch>=2.0.0
# transformers>=4.40.0
# accelerate>=0.27.0
